+++ packages/graph-explorer/src/connector/gremlin/queries/fetchSchema.test.ts
@@ -1,12 +1,16 @@
 import globalMockFetch from "@/connector/testUtils/globalMockFetch";
 import mockGremlinFetch from "@/connector/testUtils/mockGremlinFetch";
 import fetchSchema from "./fetchSchema";
+import { ClientLoggerConnector } from "@/connector/LoggerConnector";
 
 describe("Gremlin > fetchSchema", () => {
   beforeEach(globalMockFetch);
 
   it("Should return an inferred schema", async () => {
-    const schemaResponse = await fetchSchema(mockGremlinFetch());
+    const schemaResponse = await fetchSchema(
+      mockGremlinFetch(),
+      new ClientLoggerConnector()
+    );
 
     expect(schemaResponse).toMatchObject({
       vertices: [