+++ packages/graph-explorer/src/connector/openCypher/queries/fetchSchema.test.ts
@@ -1,6 +1,7 @@
 import { vi } from "vitest";
 import { SchemaResponse } from "@/connector/useGEFetchTypes";
 import fetchSchema from "./fetchSchema";
+import { ClientLoggerConnector } from "@/connector/LoggerConnector";
 
 describe("OpenCypher > fetchSchema", () => {
   it("Should return a schema", async () => {
@@ -15,7 +16,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
 
     const expected: SchemaResponse = {
       edges: [
@@ -141,7 +145,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.vertices.length).toBe(0);
   });
 
@@ -167,7 +174,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.vertices.length).toBe(0);
   });
 
@@ -193,7 +203,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.vertices.length).toBe(0);
   });
 
@@ -211,7 +224,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema).toBeDefined();
   });
 
@@ -226,7 +242,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.edges.length).toBe(0);
   });
 
@@ -252,7 +271,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.edges.length).toBe(0);
   });
 
@@ -278,7 +300,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.edges.length).toBe(0);
   });
 
@@ -304,7 +329,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema.edges.length).toBe(0);
   });
 
@@ -337,7 +365,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema).toBeDefined();
   });
 
@@ -354,7 +385,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema).toBeDefined();
   });
 
@@ -379,7 +413,10 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    const schema = await fetchSchema(openCypherFetchFn);
+    const schema = await fetchSchema(
+      openCypherFetchFn,
+      new ClientLoggerConnector()
+    );
     expect(schema).toBeDefined();
     const routeEdge = schema.edges[0];
     expect(routeEdge.total).toBeUndefined();
@@ -398,7 +435,7 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    await fetchSchema(openCypherFetchFn);
+    await fetchSchema(openCypherFetchFn, new ClientLoggerConnector());
 
     expect(openCypherFetchFn.mock.calls[3][0]).toStrictEqual(
       "MATCH () -[e:`route`]- () RETURN e AS object LIMIT 1"
@@ -420,7 +457,7 @@ describe("OpenCypher > fetchSchema", () => {
         throw new Error(query);
       });
 
-    await fetchSchema(openCypherFetchFn);
+    await fetchSchema(openCypherFetchFn, new ClientLoggerConnector());
 
     expect(openCypherFetchFn.mock.calls[3][0]).toStrictEqual(
       "MATCH () -[e:`route`]- () RETURN e AS object LIMIT 1"