+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - **Fixed** context menu styles
   ([#600](https://github.com/aws/graph-explorer/pull/600))
+- **Improved** logging on server and around schema sync
+  ([#604](https://github.com/aws/graph-explorer/pull/604))
 - **Improved** support for databases with thousands of node & edge types
   ([#599](https://github.com/aws/graph-explorer/pull/599))
 - **Fixed** alignment of close button in search panel